Level_Trainee/Part_001_Base_Syntax/7_Control_Tasks/src/main/java/ru/job4j/tictactoe/Logic3T.java
package ru.job4j.tictactoe;

import java.util.function.Predicate;

/**
 * Class Logic3T describes games logic.
 *
 * @author Alex Proshak (olexandr_proshak@ukr.net)
 */
public class Logic3T {

    /**
     * The board.
     */
    private final Figure3T[][] table;

    /**
     * The constructor with parameters.
     * @param table cells array.
     */
    public Logic3T(Figure3T[][] table) {
        this.table = table;
    }

    /**
     * The method watching whether X player win.
     * @return true in case of win.
     */
    public boolean isWinnerX() {
        return this.fillBy(Figure3T::hasMarkX, 0, 0, 1, 0)
                || this.fillBy(Figure3T::hasMarkX, 0, 1, 1, 0)
                || this.fillBy(Figure3T::hasMarkX, 0, 2, 1, 0)
                || this.fillBy(Figure3T::hasMarkX, 0, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkX, 0, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkX, 1, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkX, 2, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkX, 0, 0, 1, 1)
                || this.fillBy(Figure3T::hasMarkX, 2, 0, -1, 1);
    }

    /**
     * The method watching whether O player win.
     * @return true in case of win.
     */
    public boolean isWinnerO() {
        return this.fillBy(Figure3T::hasMarkO, 0, 0, 1, 0)
                || this.fillBy(Figure3T::hasMarkO, 0, 1, 1, 0)
                || this.fillBy(Figure3T::hasMarkO, 0, 2, 1, 0)
                || this.fillBy(Figure3T::hasMarkO, 0, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkO, 0, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkO, 1, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkO, 2, 0, 0, 1)
                || this.fillBy(Figure3T::hasMarkO, 0, 0, 1, 1)
                || this.fillBy(Figure3T::hasMarkO, 2, 0, -1, 1);
    }

    /**
     * The method checks existing available moves.
     * @return true in case of existing.
     */
    public boolean hasGap() {
        boolean result = false;
        for (int i = 0; i < this.table.length; i++) {
            for (int j = 0; j < this.table.length; j++) {
                if (!this.table[i][j].hasMarkX() && !this.table[i][j].hasMarkO()) {
                    result = true;
                    break;
                }
            }
        }
        return result;
    }

    /**
     * The method fills by player's figure.
     * @param predicate condition.
     * @param startX coordinate.
     * @param startY coordinate.
     * @param deltaX coordinate.
     * @param deltaY coordinate.
     * @return true if possible.
     */
    public boolean fillBy(Predicate<Figure3T> predicate, int startX, int startY, int deltaX, int deltaY) {
        boolean result = true;
        for (int index = 0; index != this.table.length; index++) {
            Figure3T cell = this.table[startX][startY];
            startX += deltaX;
            startY += deltaY;
            if (!predicate.test(cell)) {
                result = false;
                break;
            }
        }
        return result;
    }
}